requirements.txt
numpy~=1.19.5
matplotlib~=3.4.2
pandas~=1.3.0
scikit-learn~=0.24.2
opencv=4.5.2
tensorflow-gpu~=2.5.0
